labs/LAB02_submission/inference_api/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn>=0.22.0
pillow>=9.0.0
python-multipart>=0.0.6
